+++ src/modules/bands/schemas/bands.graphql
@@ -1,5 +1,5 @@
 type Band {
-  _id: ID!
+  id: ID!
   name: String
   origin: String
   members: [Member]
